+++ test/support/helper_case.ex
@@ -14,7 +14,6 @@ defmodule AtomTweaksWeb.HelperCase do
       import Floki, only: [attribute: 2, find: 2, text: 1], warn: false
 
       import Support.HTML
-      import Test.Helpers
     end
   end
